WIP: Refactor interfaces
